requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
decorator==5.1.0
elasticsearch==7.15.0
execnet==1.9.0
Faker~=8.8.1
greenlet==1.1.1
gssapi==1.7.0
idna==3.2
iniconfig==1.1.1
jsonpath==0.82
more-itertools==8.10.0
multi-key-dict==2.0.3
packaging==21.0
pbr==5.6.0
phoenixdb~=1.0.1
pluggy==1.0.0
protobuf==3.18.0
py==1.10.0
PyHamcrest==2.0.2
PyMySQL~=0.9.3
pyparsing==2.4.7
pytest==6.2.5
pytest-forked==1.3.0
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.4.0
python-dateutil~=2.8.1
python-jenkins==1.7.0
PyYAML==5.4.1
redis==3.5.3
requests~=2.22.0
requests-gssapi==1.2.3
requests-toolbelt==0.9.1
simplejson~=3.17.2
six==1.16.0
SQLAlchemy==1.4.25
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
Flask~=2.0.2
config~=0.5.0.post0
pymssql~=2.2.1
configparser~=5.0.0
xmltodict~=0.13.0
pandas~=1.3.5
pyodbc~=4.0.39
psycopg2~=2.9.5
chardet~=3.0.4